requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.1
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.7
idna==2.6
Markdown==2.6.11
pytz==2017.3
ranking==0.3.1
requests==2.18.4
urllib3==1.22
